requirements.txt
astroid==2.2.5
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
cloudinary==1.15.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
Django==2.1
django-autoslug-iplweb==1.9.4
django-cors-middleware==1.2.0
django-environ==0.4.5
django-extensions==2.1.6
django-mailer==1.2.6
django-rest-swagger==2.1.2
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
entrypoints==0.3
flake8==3.7.7
gunicorn==19.9.0
idna==2.8
isort==4.3.17
itypes==1.1.0
Jinja2==2.10.1
jsonfield==2.0.2
lazy-object-proxy==1.3.1
lockfile==0.12.2
MarkupSafe==1.1.1
mccabe==0.6.1
mock==2.0.0
nose==1.3.7
openapi-codec==1.3.2
pbr==5.1.3
pep8==1.7.1
Pillow==6.0.0
psycopg2==2.8.2
psycopg2-binary==2.8.2
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.4.2
pylint==2.3.1
pylint-django==2.0.8
pylint-plugin-utils==0.5
python-http-client==3.1.0
pytz==2019.1
requests==2.21.0
simplejson==3.16.0
six==1.10.0
sqlparse==0.3.0
typed-ast==1.3.4
uritemplate==3.0.0
urllib3==1.25.1
whitenoise==4.1.2
wrapt==1.11.1
